Rewrite RoomJoinPage
Renamed file to JoinRoomPage for consistency. Rewrote it to functional component (was class component before)
